school GitHub account to personal account
This commit marks the migration of the "Weather Image Classification" project from my school GitHub account to my account. Initially developed as part of a class assignment, this project has been further enhanced and expanded beyond its original scope. All previous commits and development history are maintained in the original repository. This update includes the latest improvements and features added to the project.